requirements.txt
Pillow~=7.2.0
numpy~=1.19.0
pip~=20.1.1
rsa~=4.6
pyasn1~=0.4.8
pytz~=2020.1
scipy~=1.5.1
matplotlib~=3.2.2
numba~=0.50.1
llvmlite~=0.33.0
setuptools~=47.3.1
Telethon~=1.14.0
joblib~=0.16.0
pandas~=1.0.5
python-dateutil~=2.8.1
numexpr~=2.7.1
xarray~=0.16.0
typhon~=0.8.0
netCDF4~=1.5.3
cycler~=0.10.0
docutils~=0.16
scikit-learn~=0.23.1
scikit-image~=0.17.2
imageio~=2.9.0
tifffile~=2020.7.17
seaborn~=0.10.1
networkx~=2.4
threadpoolctl~=2.1.0
six~=1.15.0
decorator~=4.4.2
pyaes~=1.6.1
pyparsing~=2.4.7
kiwisolver~=1.2.0